requirements.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
gradio
pillow
opencv-python
scikit-image